Move SwapIdentitiesFlow to confidential-identities module
