+++ node/src/main/kotlin/net/corda/node/services/CoreFlowHandlers.kt
@@ -1,9 +1,13 @@
 package net.corda.node.services
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.core.flows.*
-import net.corda.core.identity.PartyAndCertificate
+import net.corda.confidential.SwapIdentitiesFlow
+import net.corda.core.flows.AbstractStateReplacementFlow
+import net.corda.core.flows.FlowLogic
+import net.corda.core.flows.ReceiveTransactionFlow
+import net.corda.core.flows.StateReplacementException
 import net.corda.core.identity.Party
+import net.corda.core.identity.PartyAndCertificate
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.utilities.ProgressTracker
 import net.corda.core.utilities.unwrap
@@ -41,25 +45,6 @@ class NotaryChangeHandler(otherSide: Party) : AbstractStateReplacementFlow.Accep
         val isNotary = serviceHub.networkMapCache.notaryNodes.any { it.notaryIdentity == newNotary }
         if (!isNotary) {
             throw StateReplacementException("The proposed node $newNotary does not run a Notary service")
-        }
-    }
-}
-
-class SwapIdentitiesHandler(val otherSide: Party, val revocationEnabled: Boolean) : FlowLogic<Unit>() {
-    constructor(otherSide: Party) : this(otherSide, false)
-    companion object {
-        object SENDING_KEY : ProgressTracker.Step("Sending key")
-    }
-
-    override val progressTracker: ProgressTracker = ProgressTracker(SENDING_KEY)
-
-    @Suspendable
-    override fun call(): Unit {
-        val revocationEnabled = false
-        progressTracker.currentStep = SENDING_KEY
-        val legalIdentityAnonymous = serviceHub.keyManagementService.freshKeyAndCert(ourIdentity, revocationEnabled)
-        sendAndReceive<PartyAndCertificate>(otherSide, legalIdentityAnonymous).unwrap { confidentialIdentity ->
-            SwapIdentitiesFlow.validateAndRegisterIdentity(serviceHub.identityService, otherSide, confidentialIdentity)
         }
     }
 }
