+++ settings.gradle
@@ -1,6 +1,7 @@
 // The project is named 'corda-project' and not 'corda' because if this is named the same as the
 // output JAR from the capsule then the buildCordaJAR task goes into an infinite loop.
 rootProject.name = 'corda-project'
+include 'confidential-identities'
 include 'finance'
 include 'finance:isolated'
 include 'core'
